+++ tests/UnitTestEx.Api/Startup.cs
@@ -3,6 +3,10 @@ using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
+using System;
+using System.Net.Http;
+using System.Threading;
+using System.Threading.Tasks;
 
 namespace UnitTestEx.Api
 {
@@ -19,7 +23,9 @@ namespace UnitTestEx.Api
         public void ConfigureServices(IServiceCollection services)
         {
             services.AddControllers().AddNewtonsoftJson();
-            services.AddHttpClient("XXX", hc => hc.BaseAddress = new System.Uri("https://somesys"));
+            services.AddHttpClient("XXX", hc => hc.BaseAddress = new System.Uri("https://somesys"))
+                .AddHttpMessageHandler(_ => new MessageProcessingHandler())
+                .ConfigureHttpClient(hc => hc.DefaultRequestVersion = new Version(1, 2));
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
@@ -40,6 +46,17 @@ namespace UnitTestEx.Api
             {
                 endpoints.MapControllers();
             });
+        }
+
+        public class MessageProcessingHandler : DelegatingHandler
+        {
+            public static bool WasExecuted { get; set;}
+
+            protected override Task<HttpResponseMessage> SendAsync(HttpRequestMessage request, CancellationToken cancellationToken)
+            {
+                WasExecuted = true;
+                return base.SendAsync(request, cancellationToken);
+            }
         }
     }
 }