+++ tests/UnitTestEx.MSTest.Test/ProductControllerTest.cs
@@ -12,17 +12,39 @@ namespace UnitTestEx.MSTest.Test
     public class ProductControllerTest
     {
         [TestMethod]
-        public void Notfound()
+        public void Notfound_WithoutConfigurations()
         {
             var mcf = MockHttpClientFactory.Create();
-            mcf.CreateClient("XXX", new Uri("https://somesys/"))
+            mcf.CreateClient("XXX", new Uri("https://somesys"))
                 .Request(HttpMethod.Get, "products/xyz").Respond.With(HttpStatusCode.NotFound);
+
+            Startup.MessageProcessingHandler.WasExecuted = false;
 
             using var test = ApiTester.Create<Startup>();
             test.ReplaceHttpClientFactory(mcf)
                 .Controller<ProductController>()
                 .Run(c => c.Get("xyz"))
                 .AssertNotFound();
+
+            Assert.IsFalse(Startup.MessageProcessingHandler.WasExecuted);
+        }
+
+        [TestMethod]
+        public void Notfound_WithConfigurations()
+        {
+            var mcf = MockHttpClientFactory.Create();
+            mcf.CreateClient("XXX").WithConfigurations()
+                .Request(HttpMethod.Get, "products/xyz").Respond.With(HttpStatusCode.NotFound);
+
+            Startup.MessageProcessingHandler.WasExecuted = false;
+
+            using var test = ApiTester.Create<Startup>();
+            test.ReplaceHttpClientFactory(mcf)
+                .Controller<ProductController>()
+                .Run(c => c.Get("xyz"))
+                .AssertNotFound();
+
+            Assert.IsTrue(Startup.MessageProcessingHandler.WasExecuted);
         }
 
         [TestMethod]
@@ -53,6 +75,44 @@ namespace UnitTestEx.MSTest.Test
             var r = hc.GetAsync("test").Result;
             Assert.IsNotNull(r);
             Assert.AreEqual("test output", r.Content.ReadAsStringAsync().Result);
+        }
+
+        [TestMethod]
+        public void MockHttpClientFactory_NoMocking()
+        {
+            var mcf = MockHttpClientFactory.Create();
+            mcf.CreateClient("XXX").WithoutMocking();
+
+            Startup.MessageProcessingHandler.WasExecuted = false;
+
+            using var test = ApiTester.Create<Startup>();
+            var hc = test.ReplaceHttpClientFactory(mcf)
+                .Services.GetService<IHttpClientFactory>().CreateClient("XXX");
+
+            var ex = Assert.ThrowsException<AggregateException>(() => hc.GetAsync("test").Result);
+
+            Assert.IsTrue(Startup.MessageProcessingHandler.WasExecuted);
+
+            mcf.VerifyAll();
+        }
+
+        [TestMethod]
+        public void MockHttpClientFactory_NoMocking_Exclude()
+        {
+            using var mcf = MockHttpClientFactory.Create();
+            mcf.CreateClient("XXX").WithoutMocking(typeof(Startup.MessageProcessingHandler));
+
+            Startup.MessageProcessingHandler.WasExecuted = false;
+
+            using var test = ApiTester.Create<Startup>();
+            var hc = test.ReplaceHttpClientFactory(mcf)
+                .Services.GetService<IHttpClientFactory>().CreateClient("XXX");
+
+            var ex = Assert.ThrowsException<AggregateException>(() => hc.GetAsync("test").Result);
+
+            Assert.IsFalse(Startup.MessageProcessingHandler.WasExecuted);
+
+            mcf.VerifyAll();
         }
     }
 }